Implement MODINVSTOR-235: add print by default property to service point
Added an array `staffSlips` to the service point schema. This array accepts 2 properties, the `id` of the staff slip and `printByDefault` boolean property which will be used to determine the default print behavior for the slip.

Also, increased the reference data timeout to 30 seconds, since this timeout was too low for my dev machine and the builds failed consistently at 20 seconds.