MODINVSTOR-271: Unit tests can fail on referential constraint violations.
Reusing the completed Future for running the second post tenant
starts the second post tenant that inserts all the sample records
but because of the completed Future it doesn't wait until they
are inserted but immediately returns to start running the other tests.

The other tests fail because records are inserted in parallel.

Use a new (uncompleted) Future to wait until the second post tenant
has completed before starting the other unit tests.