MODINVSTOR-373: instance APIs need to set the instance HRID if not set
Implement MODINVSTOR-373 to require instance POST APIs to check the
submitted instance(s) and if the `hrid` field is not set, we generate a
new HRID and update the instance with this HRID. If the `hrid` field is
present, we preserve the submitted HRID. For the PUT API, we no longer
allow the HRID to be updated. We replace or add the `hrid` from the
existing stored instance into the submitted instance.

This logic was applied to the instance API, the batch instances API and
the synchronous batch instances API.

While creating the tests, some tests, especially the batch APIs, failed
due to a request timeout which allows operations to bleed into other
tests causing cascading failures. To minimize this, the request wait
timeout was increased on several tests. Also noticed some cleanup was
being done asynchronously without waiting for completion. Fixed that as
well. The instance storage units now complete consistently on my
over-taxed laptop.