Switch to StringUtils.isBlank()
Since we already have code that uses this library, decided to switch to
it to increase code coverage. The previous null/trim+length check would
need both sides of the or to be executed to ensure coverage. We are
assuming that Apache commons-lang3, one of the most used 3rd party Java
libraries, will work as expected.

However, it is worth noting that while we were using this library, we
did not explicitly import it in the POM file. The problem with importing
implicit dependencies is that if the explicit dependency decides to drop
or change its dependency, the implicit dependency may be unavailable,
which will cause a compile error. Fortunately, we are using several
libraries that explicitly depend on this library, so it is unlikely we'd
lose the implicit dependency. However, each library imports its own
version and it is not clear what version mod-inventory-storage will
depend on. We should *always* be explicit in our module dependencies. I
have added an explicit dependency to the latest version, as of this
commit, of `commons-lang3` in the POM.

We also have implicit dependencies to the legacy `commons-lang`. We
should move the usage of `commons-lang` to `commons-lang3` via a
tech-debt story. We should *not* add an explicit dependency to
`common-lang` in the POM.

Finally, we should consider adding dependencies checking in the POM to
ensure that we are not using implicit dependencies or depending on
libraries we no longer use. The first case can lead to stability
problems and the second case leads to deployment bloat.