Reject PUT requests that attempt to change the instance HRID
If a PUT request contains an HRID that does not exactly match the
existing `hrid` field, then the request will be requested with a 400 and
an appropriate message. While the HRID is technically case insensitive,
we are still doing an exact match for update to prevent case changes to
the existing HRID. This decision may need to be revisited in the future.