Need to handle the asynchronous nature of HRID generation
The `canPostSynchronousBatchWithSuppliedAndGeneratedHRID` test has no
guarantees on which holdings record in a batch will get a specific HRID,
only that it will get a generated HRID if not supplied.

This test assumed that the first holdings record in the batch would get
1 from the sequence. Apparently, that was true for all my local testing
and not true for the CI/CD. Now we check for either of these expected
values (there are only 2) to pass the test.