MODINVSTOR-450: Improve slow queries when migrating from Daisy to Edelweiss
This does not perform VACUUM, but a VACUUM before execution of update scripts
is highly recommended.

Optimize the query using the following strategies:
- Disable auto-updating triggers.
- Combine similar queries (same set, based on WITH clause).
- Utilize function and trigger to perform if/then conditions.

Overall performances improvements using a test database showed:
- Before these changes: ~53 minutes.
- After these changes: ~32 minutes.
* This test case is not exhaustive.
  (~3.5 million items, ~1 million instances, most other tables empty.)

This does not optimize by disabling indexes, see notes below for reasons.

Explicitly cast manually created jsonb in `renameModesOfIssuance.sql`.
This avoids warnings and any possible problem implied by the warnings.

Disabling the auto-updating triggers is helpful because the disabled tasks do
not need to be performed during this manual update. The data is already assured.
If the data is not assured, this update will therefore do so already.
This saves notable time. See examples below.

Combining similar queries saves the expensive of running multiple loops.
This only works for cases when the WHERE clauses are identical.
Multiple jsonb columns are being updated.
The concat operator is useful in this regards.

Some of the queries can be reduced if IF/THEN conditions are allowed in SQL.
In the case of PGSQL functions, IF/THEN conditions are allowed.
Temporarily create a PGSQL function to effectively combine two similar queries
into a single query. This reduces the number of loops.

In early investigations, I attempted to disable indexes during the update
process. This process worked reasonably well and showed an ~50% reduction in
time spent. The problem is that the gains were inconsistent due to unclear
reasons and the indexes eventually do need to get updated (via a
background-process, a SELECT, or a manual REINDEX). The suspected cause of the
difference is that when the database is very dirty, queries perform 50% slower.
By running a VACUUM before calling the upgrade scripts, the database can be
put into a clean state that improves performance of all queries. This has
an initial cost of the time it takes to perform a vacuum. This vacuum may
take anywhere from 1 minute to 30 minutes!

For reference, the code for disabling/enabling indexes looks like:
```
UPDATE pg_index SET indisready = false
WHERE indrelid = '${myuniversity}_${mymodule}.item'::regclass::oid;

UPDATE pg_index SET indisready = true
WHERE indrelid = '${myuniversity}_${mymodule}.item'::regclass::oid;
```

Example With Triggers Enabled:
```
--------------------------------------------------------------------------------------------------------------------------------------
 Update on item it  (cost=2.27..3687419.57 rows=3500000 width=212) (actual time=723773.371..723773.371 rows=0 loops=1)
   ->  Hash Join  (cost=2.27..3687419.57 rows=3500000 width=212) (actual time=6.378..137812.680 rows=3500000 loops=1)
         Hash Cond: (it.holdingsrecordid = hr.id)
         ->  Seq Scan on item it  (cost=0.00..3534985.00 rows=3500000 width=1753) (actual time=4.070..71617.172 rows=3500000 loops=1)
         ->  Hash  (cost=2.12..2.12 rows=12 width=54) (actual time=1.183..1.183 rows=12 loops=1)
               Buckets: 1024  Batches: 1  Memory Usage: 18kB
               ->  Seq Scan on holdings_record hr  (cost=0.00..2.12 rows=12 width=54) (actual time=1.138..1.160 rows=12 loops=1)
 Planning time: 29.087 ms
 Trigger set_id_in_jsonb: time=35607.609 calls=3500000
 Trigger set_item_md_json_trigger: time=48997.533 calls=3500000
 Trigger update_effective_location: time=14846.184 calls=3500000
 Trigger update_item_references: time=35915.253 calls=3500000
 Trigger update_item_status_date: time=31638.725 calls=3500000
 Execution time: 723777.654 ms
(14 rows)
Time: 723816.910 ms (12:03.817)
```

Example With Triggers Disabled:
```
--------------------------------------------------------------------------------------------------------------------------------------
 Update on item it  (cost=2.27..4609007.56 rows=4370069 width=212) (actual time=627148.629..627148.629 rows=0 loops=1)
   ->  Hash Join  (cost=2.27..4609007.56 rows=4370069 width=212) (actual time=0.531..159966.331 rows=3500000 loops=1)
         Hash Cond: (it.holdingsrecordid = hr.id)
         ->  Seq Scan on item it  (cost=0.00..4418679.69 rows=4370069 width=1753) (actual time=0.013..96735.732 rows=3500000 loops=1)
         ->  Hash  (cost=2.12..2.12 rows=12 width=54) (actual time=0.488..0.488 rows=12 loops=1)
               Buckets: 1024  Batches: 1  Memory Usage: 18kB
               ->  Seq Scan on holdings_record hr  (cost=0.00..2.12 rows=12 width=54) (actual time=0.477..0.483 rows=12 loops=1)
 Planning time: 0.270 ms
 Execution time: 627151.287 ms
(9 rows)
Time: 627151.896 ms (10:27.152)
```