MODINVSTOR-450: uncombine some queries due to version differences
The differences in versions as defined in schema.json prevent many of the
queries from being combined.

The `migrateItemCopyNumberToSingleValue.sql` trigger is maintained but the
update row is used like a stub to trigger the function. As a result, the
update is setting the id to the current id. The temporary trigger does the
real work.

Automatically disble triggers currently defined instead of using hardcoded
trigger names.