Speculatively add an `id` to the `illPolicy` virtual field
This branch is, until now, failing tests for obscure reasons that seem
to be to do with a pre-existing `IllPolicy` class generated from
something else in the RAMLs or (more likely) JSON Schemas.

The thing is, none of the schemas has a field called `illPolicy`, and
in fact the only use of that symbol in any of the JSON Schemas other
than in the virtual field I added is that in `illpolicies.json`, a
field called `illPolicies` has `"id": "illPolicy"`.

It turns out that no-one knows what the `id` field here means. In the
JSON Schema specifications it's used to alter resolution scope, bit
like the `<base>` element in HTML:
https://json-schema.org/draft-04/json-schema-core.html#rfc.section.7.2.2
But whatever RMB is doing with it, it ain't that. The RMB
documentation doesn't seem to mention it, and neither @adamdickmeiss
(the current maintainer of RMB) nor @nielserik (who wrote the schema
in question) knows what the `id` means.

So I am _guessing_ that it gets used as a class name -- a bit like the
way `javaType` affects things, if I have properly understood that --
and that this is why my own `illPolicy` field is colliding with
Niels-Erik's `illPolicies` field.

If that's correct, then I should be able to similarly change the
classname generated for my field by giving it an `id` of its own --
which should make the tests work, and allow me to merge this harmless
little virtual field into master.

The simplest way for me to find out is to push it and see what Jenkins
thinks.