Add `illPolicy` link-field annotation for mod-graphql (#425)
This is needed for ReShare issue PR-611.

* Bump the minor version of the holdings-storage interface (Thanks to @bohdan-suprun for the reminder to do this.)
* Speculatively add an `id` to the `illPolicy` virtual field

Until I added that `id`, this branch was failing tests for obscure reasons that seem to be to do with a pre-existing `IllPolicy` class generated from something else in the RAMLs or (more likely) JSON Schemas.

The thing is, none of the schemas has a field called `illPolicy`, and in fact the only use of that symbol in any of the JSON Schemas other than in the virtual field I added is that in `illpolicies.json`, a field called `illPolicies` has `"id": "illPolicy"`.

It turns out that no-one knows what the `id` field here means. In the JSON Schema specifications it's used to alter resolution scope, a bit like the `<base>` element in HTML:
https://json-schema.org/draft-04/json-schema-core.html#rfc.section.7.2.2
But whatever RMB is doing with it, it ain't that. The RMB documentation doesn't seem to mention it, and neither @adamdickmeiss (the current maintainer of RMB) nor @nielserik (who wrote the schema in question) knows what the `id` means.

So I am _guessing_ that it gets used as a class name -- a bit like the way `javaType` affects things, if I have properly understood that -- and that this is why my own `illPolicy` field is colliding with Niels-Erik's `illPolicies` field.

If that's correct, then I should be able to similarly change the classname generated for my field by giving it an `id` of its own -- which should make the tests work, and allow me to merge this harmless little virtual field into master.

And indeed this seems to be the case ... though no-one really knows why.