MODINVSTOR-588: New item status: Unknown
It's suggested to extend the enumeration of Item.status with a new option
"Unknown". With a shared index, circulation may be handled by the
participating libraries individually, outside of FOLIO. When presenting
the bibliographic data and their holdings from the shared index it is thus
not necessarily possible to know the status of any given item. There's not
currently a status option that indicates that the item's status isn't currently
known and item status is a required field. Having that option is convenient
if FOLIO Stripes module ui-inventory is being used for displaying records of
the shared index, since it would have to display an arbitrary (misleading)
status in lieu of 'Unknown', 'NA' orsimilar.