MODINVSTOR-508 Item status date missing when set as default value Available
Addressed this by making the item post method check for active status,
and fill it in with the item creation date from the item metadata.The status
check may not be strictly necessary, from what I can see, new item records
are automatically given a status of "Active," and you can't change that.
I'm assuming that the method I'm altering is only used for creating new item
records-it looks like the "put" method is used for updating item data.
I also had to modify several tests that expected the initial status of a
created item to be null.

refs: https://issues.folio.org/browse/MODINVSTOR-508