MODINVSTOR-509: Instance. New Instance. Instance status date is missing
In line with MODINVSTOR-508, I changed operations here to use the current
date/time as the initial status updated date.  I felt again that it was
important that creation operations behave consistently.

Getting the current date and time and inserting it into the instance creation
process was the easy part.  What took more time was rewriting tests that
either expected a null value, or were not checking the status update field
for any value at all.  It also took a while for me to realize there were two
separate endpoints that batch-create instances.  For single instance and regular
batch-creation processes, I built checking for the status update date into existing
tests where it seemed to make sense, usually into the initial check to see if such
items could be correctly created.  For batch asynchronous creation, it was actually
easier to create a separate test just for checking the status date.

refs: https://issues.folio.org/browse/MODINVSTOR-509