MODINVSTOR-637: Add disclosure file.
Holdings has circulationNotes which allows for "source.personal".
The source.personal appears to allow for firstName and lastName at the very list.

The instance contributors may have middle names in them.

There are several structures with "notes" that are free form text.
One example is a holdings record.

The userIds are used and stored.