Use Future<PutData>, not PutData method parameter
Passing the PutData variable as a reference and a method parameter
to return the result has been rejected in the code review.

The existing PutData variable must be effectively final because it
is used in other compose lambdas. Therefore the Future<PutData>
values must be copied over to the existing PutData variable.