MODINVSTOR-908: Single HRID SQL query for batch (instance/holding/item)
A batch of 10 instances currently requires 10 SQL querys fetching
a single hrid.

Combining them into one SQL is faster:

Tested with the POST /instance-storage/batch/synchronous API and
a batch size of 10 instances.

The API response time on a system without load reduces
from 88 ms to 75 ms, this is -15 %.

The throughput on a system under load (many parallel requests to
this API) increases from 46 requests/second to 64 request/second
(= from 460 instance/second to 640 instances/second),
this is +39 %.