MODINVSTOR-923: Reject batch upsert changing HRID
Changing the HRID is rejected in the single record
(instance, holding, item) PUT APIs.

However, the synchronous batch API allow it. This is a bug.

The batch API Java code doesn't fetch the records before running
the update query for performance reasons. Therefore we need to
check the HRID in the database.

The new trigger method hrid_lock() fails the update on HRID change.
It replaces HridValidators.java that is removed.

A custom PostgreSQL error code 23578 indicates the hrid lock violation.

Currently no other module relies on the error message:
https://github.com/search?q=org%3Afolio-org+%22hrid+field+cannot%22&type=code