MODINVSTOR-977: hasRemoveEvent fixing sporadic unit test failures
About 30% of all builds fail because of at least one of these failures:

```
InstanceStorageTest.canDeleteInstancesByCql:487
  Expected: is "DELETE"
       but: was "CREATE"
```

```
  HoldingsStorageTest.canDeleteHoldingsByCql:647
Expected: is "DELETE"
     but: was "CREATE"
```

```
ItemStorageTest.canDeleteItemsByCql:2045
  Expected: is "DELETE"
       but: was "CREATE"
```

The current tests assert that the last event is the expected DELETE event. Due to the async nature of Kafka messages there might be a CREATE event.

Rewrite the test so that it succeeds if the expected DELETE event exists at any position in the event stream.

Write a hasRemoveEvent method that can be used with await().until(() -> hasRemoveEvent(...));