MODINVSTOR-982: Kafka events are not consistently cleared between tests.
The states are not always cleaned before tests.
This results in tests not actually running in isolation.
This can cause inconsistent and incorrect test results.

This takes a compatible approach by implementing a custom function `removeAllEvents()`.
Ideally, this should instead be named `beforeEach()` but then any class that implements this would need to be updated to call `super.removeAllEvents()`.