MODINVSTOR-982: The FakeKafkaConsumer should return first and last events by date.
The events arrive in an asynchronous manner.
This results in the possibility of the first and last events being ordered differently than the date those events were generated.
If this list is not supposed to be ordered by date such that fetching the last item is the last item by date and the first item is the first item by date, then this commit should be reverted.

These changes are independent of "MODINVSTOR-977: hasRemoveEvent fixing sporadic unit test failures (#835)" (8a3327103dce753720d5b026a01b715d44b0191f).