MODINVSTOR-982: Double down on test suite design, improving isolation.
There are several problems with the structure that results in more confusing code.
The order in which the tests are executing is a big factor.
There are several test classes that are not actually included when they should be.

This is a re-design that double-downs on the test suite paradigm.
This allows for better control over how the tests are run.
This should make it easier to upgrade to Junit5.
This should make it easier to find and fix the design problems with the tests at a more leisurely pace.

This has a cost of now requiring all tests intended to be run needing to be manually added to a test suite class.
I would also note that it seems `InstallUpgradeIt.java` has never been run before (is it run through some other means?).

Many of the changes here are done in a minimalistic fashion.
There is a lot of room for improvement and these changes are considered good enough for this pass.
I expect that the migration to Junit5 is the better time to make better and more sweeping changes.

The `KafkaTopicsExistsTest` class is not under the `StorageTestSuite` but depends on Kafka (and related) started by it.
This is an isolation problem.
As a short term solution, have that class start and stop its own instances.
Future designs should change or improve upon this behavior.

Summary:
- Update the `pom.xml` to use the `AllTestSuite.java` class.
- Create test suites for all tests.
- Extract the Kafka and Vertx into their own utility classes.
- Ensure proper setup and cleanup of Kafka and Vertx.
- Fix imports structure when already having to update the imports (removing wildcards, etc..).
- Use static imports for static methods more often to simplify the code slightly.
- For tests not under `api/` directory that use Kafka now start and stop their own instances.