MODINVSTOR-982: Always commit Kafka, add timeout to block, and utilize SneakyThrows
Utilize `@SneakyThrows` annotation on all before/after methods using junit annotations.
This should make overriding the methods easier and cleaner if necessary.

Add `removeAllEvents(false)` to all methods using `@Before` annotation to ensure anything they do does not result in data in the Kafka message arrays.

Rewrite the `removeAllEvents()` method to utilize a timeout.
The `commit()` with a blocking `get()` does appear to improve consistency of test success.
This has a huge cost in time because tests are now taking as long as 10 minutes.
Explicitly reduce this to a 300 millisecond timeout and don't care if it times out (hide exception).
A debug logger is used in case one wants to identify these during testing.
This reduces the time tests take to about about 7 minutes.

Always have Kafka commit even when not using the timeout.
Add a commit that neither waits nor blocks when not asking to block.