MODINVSTOR-982: Change commit() design, disable autocommit, add timestamp
Added commits in the middle of some tests.
This seemed to result in more consistent test failures.

Added a timestamp support in Kafka to prevent and log any stale or out of date messages.
This appears to show no results.
This suggests that the inconsistent problems are not in the event part of the process.

The commits in the middle of some tests revealed that the commit process is likely causing the problem.
I may have had the problem entirely backwards.
Do not do more commits, instead do less.

Remove much of the commit logic and maintain the clear arrays logic.
Rename the FakeKafkaConsumer removeAllEvents function to clearAllEvents to avoid potential confusion with the TestBase removeAllEvents function.

Disable autocommit to help prevent the random commits from breaking things.
The existing manual commit functions are being left in case they are needed in special cases.