MODINVSTOR-982: Use AfterAll() rather than After() for ModesOfIssuanceMigrationScriptTest
When data is not cleared the instance storage tests fail with:
```
14:52:21 [] [test_tenant] [] [mod_inventory_storage] INFO  RestRouting          invoking postInstanceStorageBatchInstances
14:52:21 [] [test_tenant] [] [mod_inventory_storage] ERROR tanceStorageBatchAPI Failed to generate an instance HRID
io.vertx.pgclient.PgException: ERROR: insert or update on table "instance" violates foreign key constraint "modeofissuanceid_mode_of_issuance_fkey" (23503)
...
```

To prevent this the data must be cleared after the class.
The previous behavior is clearing that both before and after each test.
Performing this after each test is unnecessary.
Change the `@After` to `@AfterClass`.