MODINVSTOR-982: Partially revert excessive Suite usage
It has been decided to rename the files themselves to avoid excessive Suite usage.
As separate commit and pull request are to be used to better isolate changes.

This partially reverts the changes regarding adding everything to a Suite.
The code will be changed anyway so there is no point in added something to then immediately remove it.