fix(response-message): Make response message more informative.
Previously, during POST on the /instance-storage/instances with instance with existing hrid, response body contained database-related info "lower(f_unaccent(jsonb ->> 'hrid'::text))" and processed as success to the handler in the postInstanceStorageInstances method. This commit solves this problem by creating handler to process error message and make it informative for user.

- Add handler in InstanceStorageApi class.
- Add handleResponse, extractErrorMessage methods in EndpointHandler class.

Closes MODINVSTOR-1100