MODINVSTOR-1423: Show how many holdings records need sourceId migration (#1196)
* MODINVSTOR-1423: Show how many holdings records need sourceId migration

https://folio-org.atlassian.net/browse/MODINVSTOR-1423

Add ramsons-sourceid-migration-analyze.sql to mod-inventory-storage git
repository.

Before migration to Ramsons a Sysop need to know how many holdings records lack the sourceId for each tenant.

This is needed to decide whether to run migration script, and to estimate how long the migration script may run.

Adding the SQL script to the mod-inventory-storage git repository allows for automated download and versioning.

The script must use the sourceid field that has a database index, it must not use the jsonb->>'sourceId' property that lacks a database index.

* Add NEWS

* Move script to migration directory